+++ src/pages/Main.js
@@ -14,9 +14,9 @@ import ErrorWrapper from 'src/components/ErrorWrapper';
 import Favicon from 'src/components/Favicon';
 import FirecloudNotification from 'src/components/FirecloudNotification';
 import IdleStatusMonitor from 'src/components/IdleStatusMonitor';
-import ImportStatus from 'src/components/ImportStatus';
 import SupportRequest from 'src/components/SupportRequest';
 import { TitleManager } from 'src/components/TitleManager';
+import ImportStatus from 'src/data/ImportStatus';
 import { getOidcConfig } from 'src/libs/auth';
 import { getEnabledBrand } from 'src/libs/brand-utils';
 import { PageViewReporter } from 'src/libs/events';
